requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
dependency-injector==4.41.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
multimethod==1.10
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pandas-stubs==2.2.1.240316
pandera==0.18.3
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
typeguard==4.2.1
types-pytz==2024.1.0.20240203
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
wrapt==1.16.0
yarl==1.9.4
